zero_weighted_pca_model
Built a zero_weighted_pca_model() function that will take data, wavenumbers, and zero weight range(s) in the form of tuple(s). It will then build a pca model that ensures the contribution of data in the specified tuple ranges is zero'd. This should make any pca smoothing with this function also make the values in those ranges zero (and should be removed after)