core_mask_cleaner and identify_cores_and_plot functions
Creation of two functions that can be used together as part of core finding from a tissue TMA:

core_mask_cleaner can take a core mask that is a little noisy/has debris in and cleans up the mask to hopefully speed up core finding.

identify_cores_and_plot identifies cores within the TMA, collecting the [y, x] coordinates of the grid to isolate each core, along with a labelled image to identify which grid corresponds to which core.